implemented form validation

+++ views/upload.php
@@ -5,29 +5,38 @@ echo renderUpload();
 
 $formMessage = '';
 if(isset($_POST['singlefileupload'])) {
-    $fileHandle = processUpload($_FILES['userfile']);
+    $formMessage = formValidation($_POST);
 
-    if($fileHandle['status']){
-        $resizeImage = img_resize($fileHandle['filename'], $fileHandle['filename'], 600, 600, $quality=100);
+    if ($formMessage != '') {
+        $formMessage;
+    }
 
-        if($resizeImage['status']){
-            $thumbName = $config['thumb_dir']. 'thumb' . basename($fileHandle['filename']);
-            $createThumb = img_resize($fileHandle['filename'], $thumbName,150,150);
+    else {
+        $fileHandle = processUpload($_FILES['userfile']);
 
-            if($createThumb['status']){
-                $dbTransaction = addToDB(basename($fileHandle['filename']),htmlentities($_POST['title']),htmlentities($_POST['description']),$fileHandle['filename'],$thumbName,$resizeImage['newWidth'],$resizeImage['newHeight']);
-                $formMessage = $createThumb['msg'];
-                if(!$dbTransaction){
-                    unlink($fileHandle['filename']);
-                    unlink($thumbName);
-                    $formMessage = "DB transaction unsuccessful, deleted file and thumbnail.";
+        if($fileHandle['status']){
+            $resizeImage = img_resize($fileHandle['filename'], $fileHandle['filename'], 600, 600, $quality=100);
 
+            if($resizeImage['status']){
+                $thumbName = $config['thumb_dir']. 'thumb' . basename($fileHandle['filename']);
+                $createThumb = img_resize($fileHandle['filename'], $thumbName,150,150);
+
+                if($createThumb['status']){
+                    $dbTransaction = addToDB(basename($fileHandle['filename']),htmlentities($_POST['title']),htmlentities($_POST['description']),$fileHandle['filename'],$thumbName,$resizeImage['newWidth'],$resizeImage['newHeight']);
+                    $formMessage = $createThumb['msg'];
+                    if(!$dbTransaction){
+                        unlink($fileHandle['filename']);
+                        unlink($thumbName);
+                        $formMessage = "DB transaction unsuccessful, deleted file and thumbnail.";
+
+                    }
                 }
             }
+        } else {
+            $formMessage = $fileHandle['msg'];
         }
-    } else {
-        $formMessage = $fileHandle['msg'];
     }
+
 }
 $formTemplate = file_get_contents($config['app_dir'].'/templates/form.html');
 echo renderForm($formTemplate,$formMessage);
